Implement action

+++ migrations/1761920249043_add-initial-tables.sql
@@ -6,7 +6,7 @@ CREATE TABLE players (
   id SERIAL PRIMARY KEY,
   player_id VARCHAR(255) NOT NULL UNIQUE,
   player_name VARCHAR(255),
-  score NUMERIC NOT NULL,
+  score INTEGER NOT NULL,
   created_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
